Make lastruntime a double
So it matches other time variables. It makes sure there will not be two INTERIM
reports in the same second (making the database guys happy).

+++ src/curlops.h
@@ -13,7 +13,7 @@
 
 struct myprogress {
   int stream;
-  long lastruntime;
+  double lastruntime;
   double lastdlbytes;
   CURL *curl;
   struct {
